Update slideshow in ParagraphFormat.
