+++ lib/helpers/paragraphFormat.dart
@@ -1,6 +1,5 @@
 import 'dart:math' as math;
 
-import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -9,6 +8,7 @@ import 'package:readr_app/helpers/constants.dart';
 import 'package:readr_app/models/contentList.dart';
 import 'package:readr_app/models/paragraph.dart';
 import 'package:readr_app/widgets/embeddedCodeWidget.dart';
+import 'package:readr_app/widgets/imageAndDescriptionSlideShowWidget.dart';
 import 'package:readr_app/widgets/imageDescriptionWidget.dart';
 import 'package:readr_app/widgets/mMVideoPlayer.dart';
 
@@ -80,8 +80,7 @@ class ParagraphFormat {
       }
       break;
       case 'slideshow': {
-        var width = MediaQuery.of(context).size.width-32;
-        return buildSlideshowWidget(paragraph.contents, width);
+        return ImageAndDescriptionSlideShowWidget(contentList: paragraph.contents);
       }
       break;
       case 'youtube': {
@@ -189,81 +188,6 @@ class ParagraphFormat {
           ],
         );
       }
-    );
-  }
-
-  Widget buildSlideshowWidget(ContentList contentList, double width) {
-    double theSmallestRatio;
-    contentList.forEach(
-      (content) { 
-        if(theSmallestRatio == null) {
-          theSmallestRatio = content.aspectRatio;
-        }
-        else if(theSmallestRatio > content.aspectRatio) {
-          theSmallestRatio = content.aspectRatio;
-        }
-      }
-    );
-    double carouselRatio = theSmallestRatio * 0.85;
-    CarouselOptions options = CarouselOptions(
-      viewportFraction: 1,
-      aspectRatio: carouselRatio,
-      enlargeCenterPage: true,
-      onPageChanged: (index, reason) {},
-    );
-    CarouselController carouselController = CarouselController();
-    List<Widget> silders = contentList.map(
-      (content) => ImageDescriptionWidget(
-        imageUrl: content.data, 
-        description: content.description, 
-        width: width,
-        aspectRatio: content.aspectRatio,
-      )
-    ).toList();
-
-    return Stack(
-      children: [
-        CarouselSlider(
-          items: silders,
-          carouselController: carouselController,
-          options: options,
-        ),
-        
-        Align(
-          alignment: Alignment.centerLeft,
-          child: InkWell(
-            child: SizedBox(
-              width: width * 0.1,
-              height: width/carouselRatio,
-              child: Icon(
-                Icons.arrow_back_ios,
-                color: appColor,
-                size: 36,
-              ),
-            ),
-            onTap: () {
-              carouselController.previousPage();
-            },
-          ),
-        ),
-        Align(
-          alignment: Alignment.centerRight,
-          child: InkWell(
-            child: SizedBox(
-              width: width * 0.1,
-              height: width/carouselRatio,
-              child: Icon(
-                Icons.arrow_forward_ios,
-                color: appColor,
-                size: 36,
-              ),
-            ),
-            onTap: () {
-              carouselController.nextPage();
-            },
-          ),
-        ),
-      ],
     );
   }
